Use `print` instead of `puts` and Add idx by one
because `puts` append new line at the end of each line, we don't get
the desired board.
if there is `no symbol` in `grid`, the index will puted in place of the
`symbol`. but if we don`t add `one` to the index the numbers on the
board will be start from `0` and end at `8` which is false
representation. So, we shoul add `one` to theire index.